Create python advacne web browser
To use this script, you will need to install the google-auth and google-api-python-client libraries and set the TASK_LIST_ID, TASK_NAME, and TASK_NOTES variables to the appropriate values. You will also need to follow the instructions in the Google Tasks API documentation to set up your project and authorize the script to access your Google Tasks.